+++ lib_xcore_math/src/arch/ref/vect_inverse.c
@@ -1,16 +1,12 @@
-// Copyright 2020-2022 XMOS LIMITED.
+// Copyright 2020-2023 XMOS LIMITED.
 // This Software is subject to the terms of the XMOS Public Licence: Version 1.
 
 #include <stdint.h>
 #include <stdio.h>
 
 #include "xmath/xmath.h"
-#include "../../vect/vpu_helper.h"
+#include "vpu_helper.h"
 #include "xmath/xs3/vpu_scalar_ops.h"
-
-
-
-
 
 
 void vect_s16_inverse(
@@ -20,14 +16,10 @@ void vect_s16_inverse(
     const unsigned scale)
 {
     const int32_t dividend = 1 << scale;
-    for(int k = 0; k < length; k++){
-        a[k] = (dividend / b[k]);
+    for(unsigned k = 0; k < length; k++){
+        a[k] = (int16_t) (dividend / b[k]);
     }
 }
-
-
-
-
 
 headroom_t vect_s32_inverse(
     int32_t a[],
@@ -38,8 +30,8 @@ headroom_t vect_s32_inverse(
 
     const int64_t d = (0x1LL << scale);
 
-    for(int k = 0; k < length; k++){
-        a[k] = d / b[k];
+    for(unsigned k = 0; k < length; k++){
+        a[k] = (int32_t) (d / b[k]);
         
         // printf("0x%016llX / %ld = %ld\n", (uint64_t) d, b[k], a[k]);
     }
