+++ test/vect_tests/src/vect/float/test_vect_complex_f32_mul.c
@@ -63,18 +63,18 @@ TEST(vect_complex_f32_mul, vect_complex_f32_mul)
   DWORD_ALIGNED
   complex_float_t c[MAX_LEN];
 
-  for(int v = 0; v < REPS; v++){
+  for(unsigned int v = 0; v < REPS; v++){
     const unsigned old_seed = seed;
 
     unsigned len = pseudo_rand_uint(&seed, 1, MAX_LEN-1);
-    setExtraInfo_RSL(v, old_seed, len);  
+    setExtraInfo_RSL(v, old_seed, len);
     // printf("0x%08X\n", old_seed);
 
-    for(int k = 0; k < len; k++){
-      b[k].re = ldexpf(pseudo_rand_int32(&seed), pseudo_rand_int(&seed, -40, 10));
-      b[k].im = ldexpf(pseudo_rand_int32(&seed), pseudo_rand_int(&seed, -40, 10));
-      c[k].re = ldexpf(pseudo_rand_int32(&seed), pseudo_rand_int(&seed, -40, 10));
-      c[k].im = ldexpf(pseudo_rand_int32(&seed), pseudo_rand_int(&seed, -40, 10));
+    for(unsigned int k = 0; k < len; k++){
+      b[k].re = ldexpf((float) pseudo_rand_int32(&seed), pseudo_rand_int(&seed, -40, 10));
+      b[k].im = ldexpf((float) pseudo_rand_int32(&seed), pseudo_rand_int(&seed, -40, 10));
+      c[k].re = ldexpf((float) pseudo_rand_int32(&seed), pseudo_rand_int(&seed, -40, 10));
+      c[k].im = ldexpf((float) pseudo_rand_int32(&seed), pseudo_rand_int(&seed, -40, 10));
 
       double RR = ((double)b[k].re) * ((double)c[k].re);
       double RI = ((double)b[k].re) * ((double)c[k].im);
@@ -90,14 +90,14 @@ TEST(vect_complex_f32_mul, vect_complex_f32_mul)
         continue;
       }
       
-      expected[k].re = RR - II;
-      expected[k].im = IR + RI;
+      expected[k].re = (float) (RR - II);
+      expected[k].im = (float) (IR + RI);
 
     }
     
     vect_complex_f32_mul(a, b, c, len);
 
-    for(int k = 0; k < len; k++){
+    for(unsigned int k = 0; k < len; k++){
 
       complex_float_t thresh;
       thresh.re = 16*get_lsb_value(expected[k].re);
@@ -127,18 +127,18 @@ TEST(vect_complex_f32_mul, vect_complex_f32_conj_mul)
   DWORD_ALIGNED
   complex_float_t c[MAX_LEN];
 
-  for(int v = 0; v < REPS; v++){
+  for(unsigned int v = 0; v < REPS; v++){
     const unsigned old_seed = seed;
 
     unsigned len = pseudo_rand_uint(&seed, 1, MAX_LEN-1);
-    setExtraInfo_RSL(v, old_seed, len);  
+    setExtraInfo_RSL(v, old_seed, len);
     // printf("0x%08X\n", old_seed);
 
-    for(int k = 0; k < len; k++){
-      b[k].re = ldexpf(pseudo_rand_int32(&seed), pseudo_rand_int(&seed, -40, 10));
-      b[k].im = ldexpf(pseudo_rand_int32(&seed), pseudo_rand_int(&seed, -40, 10));
-      c[k].re = ldexpf(pseudo_rand_int32(&seed), pseudo_rand_int(&seed, -40, 10));
-      c[k].im = ldexpf(pseudo_rand_int32(&seed), pseudo_rand_int(&seed, -40, 10));
+    for(unsigned int k = 0; k < len; k++){
+      b[k].re = ldexpf((float) pseudo_rand_int32(&seed), pseudo_rand_int(&seed, -40, 10));
+      b[k].im = ldexpf((float) pseudo_rand_int32(&seed), pseudo_rand_int(&seed, -40, 10));
+      c[k].re = ldexpf((float) pseudo_rand_int32(&seed), pseudo_rand_int(&seed, -40, 10));
+      c[k].im = ldexpf((float) pseudo_rand_int32(&seed), pseudo_rand_int(&seed, -40, 10));
 
       double RR = ((double)b[k].re) * ((double)c[k].re);
       double RI = ((double)b[k].re) * ((double)c[k].im);
@@ -154,14 +154,14 @@ TEST(vect_complex_f32_mul, vect_complex_f32_conj_mul)
         continue;
       }
       
-      expected[k].re = RR + II;
-      expected[k].im = IR - RI;
+      expected[k].re = (float) (RR + II);
+      expected[k].im = (float) (IR - RI);
 
     }
     
     vect_complex_f32_conj_mul(a, b, c, len);
 
-    for(int k = 0; k < len; k++){
+    for(unsigned int k = 0; k < len; k++){
 
       complex_float_t thresh;
       thresh.re = 16*get_lsb_value(expected[k].re);
